Add decorators module ...
Moves the operations_mode decorator into the new module. Adds another
decorator for checking the python system path and adding any paths
specified in our designated .pth file. This applies for any file that
might need access to the hub-parsers
